refactor: context and reducer file naming
