+++ src/assets/index.js
@@ -2,6 +2,6 @@ import chessbase from "./chessbase.webp";
 import chessTalk from "./chessTalk.webp";
 import ericRosen from "./ericRosen.webp";
 import gothamChess from "./gothamChess.webp";
-import playlistImage from "./chess-dark-brown-light-pieces.webp";
+import playlistImage from "./chessPieces.webp";
 
 export { chessbase, chessTalk, ericRosen, gothamChess, playlistImage };
