+++ src/context/index.js
@@ -1,8 +1,8 @@
-export { VideosProvider, useVideos } from "./videos-context";
-export { CategoryProvider, useCategory } from "./category-context";
-export { AuthProvider, useAuth } from "./auth-context";
-export { LikesProvider, useLikes } from "./likes-context";
-export { PlaylistsProvider, usePlaylists } from "./playlists-context";
-export { PlaylistModalProvider, usePlaylistModal } from "./playlistModal-context";
-export { HistoryProvider, useHistory } from "./history-context";
-export { WatchLaterProvider, useWatchLater } from "./watchLater-context";
+export { VideosProvider, useVideos } from "./videosContext";
+export { CategoryProvider, useCategory } from "./categoryContext";
+export { AuthProvider, useAuth } from "./authContext";
+export { LikesProvider, useLikes } from "./likesContext";
+export { PlaylistsProvider, usePlaylists } from "./playlistsContext";
+export { PlaylistModalProvider, usePlaylistModal } from "./playlistModalContext";
+export { HistoryProvider, useHistory } from "./historyContext";
+export { WatchLaterProvider, useWatchLater } from "./watchLaterContext";
